clkmgr: separate clkmgr pub headers from libptpmgmt pub headers
 - move clock manager public headers into clkmgr/pub folder so that
   installation of Clock Manager can be excluded if user only needs
   libptpmgmt library.
 - make sure all headers that needed to build sample application are
   located in clkmgr/pub folder.
 - add clkmgr/pub as include directory in client, common and proxy
   compilation.
 - add clkmgr/pub folder into SRC_FILES_DIR.
 - make sure all public header files is using .h extension, not .hpp.
 - rename clkmgr_init.hpp to clockmanager.h to indicate it is the main
   public header file which user should include.
 - clockmanager.h will link to c wrapper, clkmgr/clockmanager_c.h if
   __cplusplus is not defined.
 - in future when make install is supported, clockmanager.h will be located
   in standard system directory (/usr/local/include) while all the other
   public headers will be in clkmgr folder (/usr/local/include/clkmgr).
 - remove clkmgr_ prefix for header files that inside clkmgr folder.
 - add missing header guard for clockmanager_c.h.
 - move event state and event count related structure into new pub header
   named event_state.h.
 - move useful macro into new pub header named utility.h.
 - update description of public header files.
