+++ clkmgr/client/subscribe_msg.hpp
@@ -17,9 +17,9 @@
 #define CLIENT_SUBSCRIBE_MSG_HPP
 
 #include <client/message.hpp>
+#include <clkmgr/client_state.h>
 #include <common/subscribe_msg.hpp>
 #include <rtpi/condition_variable.hpp>
-#include "clkmgr_client_state.hpp"
 
 __CLKMGR_NAMESPACE_BEGIN
