+++ clkmgr/proxy/Makefile
@@ -12,12 +12,12 @@
 CLKMGR_PROXY_DIR := $(shell pwd)
 CLKMGR_TOPLEVEL_DIR = $(CLKMGR_PROXY_DIR)/..
 CLKMGR_COMMON_DIR = $(CLKMGR_TOPLEVEL_DIR)/common
-LIBPTPMGMT_PUB_DIR = $(CLKMGR_TOPLEVEL_DIR)/../pub/
+CLKMGR_PUB_DIR = $(CLKMGR_TOPLEVEL_DIR)/pub/
 PTPMGMT_DIR := $(CLKMGR_TOPLEVEL_DIR)/../.libs
 LIBS = pthread rt ptpmgmt rtpi
 LIBS_FLAGS = $(foreach lib,$(LIBS),-l$(lib)) -L$(PTPMGMT_DIR)
 OBJ = client main message msgq_tport notification_msg null_tport connect_msg subscribe_msg transport connect_ptp4l thread
-COMMON_OBJ = print sighandler transport msgq_tport clkmgr_subscription message connect_msg subscribe_msg notification_msg
+COMMON_OBJ = print sighandler transport msgq_tport subscription message connect_msg subscribe_msg notification_msg
 OBJ_FILES = $(foreach f,$(OBJ),$(f).o)
 DEPENDS = $(foreach f,$(OBJ),$(f).d)
 COMMON_OBJ_DIR = $(CLKMGR_COMMON_DIR)/obj
@@ -36,7 +36,7 @@ build-prep:
 
 %.d : %.cpp
 	echo "[MAKEDEP]" $@
-	g++ -I $(CLKMGR_TOPLEVEL_DIR) -I $(LIBPTPMGMT_PUB_DIR) -MM -MF $@ $<
+	g++ -I $(CLKMGR_TOPLEVEL_DIR) -I $(CLKMGR_PUB_DIR) -MM -MF $@ $<
 
 .PHONY: clean
 clean:
@@ -68,5 +68,5 @@ clkmgr_proxy: $(OBJ_FILES) $(COMMON_OBJ_FILES)
 
 %.o : %.cpp
 	echo "[COMPILE]" $<
-	g++ -c $< -I $(CLKMGR_TOPLEVEL_DIR) -I $(LIBPTPMGMT_PUB_DIR) -fdiagnostics-color=always
+	g++ -c $< -I $(CLKMGR_TOPLEVEL_DIR) -I $(CLKMGR_PUB_DIR) -fdiagnostics-color=always
