+++ clkmgr/sample/Makefile
@@ -14,7 +14,7 @@ CC = gcc
 CFLAGS = -Wall
 CXX = g++
 CXXFLAGS = -Wall
-INCLUDES = -I$(shell pwd)/.. -I$(shell pwd)/../../pub
+INCLUDES = -I$(shell pwd)/.. -I$(shell pwd)/../pub
 LDFLAGS = -L$(shell pwd)/../client
 LDLIBS = -lclkmgr
 TARGET = clkmgr_test
